Split out URL building tests into separate class.
