+++ Example/Tests/ProvideIdentTests.swift
@@ -8,7 +8,6 @@
 
 import XCTest
 import provide
-@testable import provide_Example
 
 class ProvideIdentTests: XCTestCase {
     
@@ -59,34 +58,6 @@ class ProvideIdentTests: XCTestCase {
                 XCTAssertTrue(stub.mostRecentRequest!.debugDescription.contains(email))
             }
         }
-    }
-    
-    // MARK: - Helper Tests
-    
-    func testBuildUrl() {
-        performTestForBuildUrl(path: "path-to-resource")
-    }
-    
-    func testBuildUrlWithQueryString() {
-        performTestForBuildUrl(path: "path-to-resource?key=value&with=a%20space")
-    }
-    
-    func testBuildUrlWithUnescapedQueryString() {
-        let base = URL(string: "https://sub.domain.tld/")
-        XCTAssertNotNil(base)
-        let path = "path-to-resource?key=value&with=a space" // "path-to-✊🦃👀"
-        let result = ProvideApiClient().buildUrl(path: path, baseUrl: base!)
-        // TEST: update once conditional URL encoding is supported: XCTAssertEqual(result?.absoluteString, "\(base!)\(path)")
-        XCTAssertNil(result)
-    }
-    
-    // MARK: - Helper Methods
-    
-    func performTestForBuildUrl(path: String, function: String = #function, line: Int = #line) {
-        let base = URL(string: "https://sub.domain.tld/")
-        XCTAssertNotNil(base)
-        let result = ProvideApiClient().buildUrl(path: path, baseUrl: base!)
-        XCTAssertEqual(result?.absoluteString, "\(base!)\(path)")
     }
     
 }
